Remove replicates in nutrient data

Remove replicates in SWMP nutrient data to keep approximate monthly time step

Usage

rem_reps(swmpr_in, ...)

# S3 method for class 'swmpr'
rem_reps(swmpr_in, FUN = function(x) mean(x, na.rm = TRUE), ...)

Arguments

swmpr_in

input swmpr object

...

arguments passed to other methods

FUN

function to combine values, defaults to mean

Value

Returns a swmpr object for nutrient data with no replicates.

Details

Raw nutrient data obtained from the CDMO will usually include replicate samples that are taken within a few minutes of each other. This function combines nutrient data that occur on the same day. The datetimestamp column will always be averaged for replicates, but the actual observations will be combined based on the user-supplied function which defauls to the mean. Other suggested functions include the median, min, or max. The entire function call including treatment of NA values should be passed to the FUN argument (see the examples). The function is meant to be used after qaqc processing, although it works with a warning if QAQC columns are present.

See also

Examples

## get nutrient data
data(apacpnut)
swmp1 <- apacpnut

# remove replicate nutrient data
rem_reps(swmp1)
#> Warning: QAQC columns present, removed from output
#>           datetimestamp        po4f       nh4f   no2f   no3f       no23f
#> 1   2002-04-02 11:55:30 0.004000000 0.02800000 0.0020 0.0470 0.049000000
#> 2   2002-04-30 11:15:00 0.014000000 0.13800000 0.0050 0.1150 0.120000000
#> 3   2002-06-04 11:03:00 0.006000000 0.04900000 0.0020 0.0240 0.026000000
#> 4   2002-07-02 09:54:00 0.015500000 0.08800000 0.0020     NA 0.039500000
#> 5   2002-08-06 12:13:00 0.011000000 0.04000000 0.0030 0.0360 0.039000000
#> 6   2002-09-10 09:00:00 0.026000000 0.03900000 0.0030 0.0130 0.016000000
#> 7   2002-10-08 08:56:00 0.009500000 0.03650000 0.0090 0.0195 0.028500000
#> 8   2002-11-04 10:05:30 0.006500000 0.06100000 0.0055 0.0970 0.102500000
#> 9   2002-12-02 12:10:30 0.010500000 0.06100000 0.0040 0.1095 0.113500000
#> 10  2003-01-06 10:31:00 0.014000000 0.23700000 0.0030 0.2370 0.240000000
#> 11  2003-02-03 11:11:00 0.004000000 0.03750000 0.0020     NA 0.095500000
#> 12  2003-03-03 10:21:00 0.015000000 0.06150000 0.0060 0.1290 0.135000000
#> 13  2003-04-08 08:30:00 0.025000000 0.15900000 0.0050 0.1740 0.179000000
#> 14  2003-05-06 09:49:00 0.003500000 0.05700000 0.0070 0.2290 0.236000000
#> 15  2003-06-02 10:16:00 0.003500000 0.05650000 0.0090 0.2210 0.230000000
#> 16  2003-07-08 12:47:00 0.001000000 0.03400000 0.0020 0.0490 0.051000000
#> 17  2003-08-04 10:18:30 0.001000000 0.03500000 0.0030 0.2690 0.272000000
#> 18  2003-09-09 12:12:00 0.002500000 0.01400000 0.0050 0.1385 0.143500000
#> 19  2003-10-07 12:48:30 0.004500000 0.12450000 0.0020 0.0205 0.022500000
#> 20  2003-11-05 09:38:30 0.003500000 0.01700000 0.0020 0.0260 0.028500000
#> 21  2003-12-08 10:57:30 0.001500000 0.00400000 0.0020 0.1770 0.179000000
#> 22  2004-01-08 10:51:00 0.001000000 0.00650000 0.0020     NA 0.017500000
#> 23  2004-02-09 11:57:30 0.001000000 0.00500000 0.0020 0.0785 0.080500000
#> 24  2004-03-09 11:11:00 0.002000000 0.03050000 0.0030 0.2840 0.287000000
#> 25  2004-04-05 09:52:30 0.001000000 0.02050000 0.0020 0.0970 0.101000000
#> 26  2004-05-04 10:06:00 0.007000000 0.08550000 0.0045 0.0890 0.093500000
#> 27  2004-06-07 09:17:00 0.006500000 0.05950000 0.0040 0.0365 0.040500000
#> 28  2004-07-06 08:57:30 0.001000000 0.05800000 0.0060 0.0610 0.067000000
#> 29  2004-08-02 12:41:00 0.001000000 0.01000000 0.0020     NA 0.023500000
#> 30  2004-09-08 14:07:30 0.003000000 0.04650000 0.0120 0.1095 0.121500000
#> 31  2004-10-04 10:08:30          NA 0.04800000 0.0115 0.0430 0.054500000
#> 32  2004-11-08 10:16:30 0.002000000 0.01800000 0.0115 0.1810 0.192500000
#> 33  2004-11-30 12:02:00 0.007000000 0.01850000 0.0050 0.0380 0.043000000
#> 34  2005-01-04 10:47:30 0.004000000 0.02750000 0.0020 0.0140 0.017500000
#> 35  2005-02-07 11:18:00 0.001000000 0.01850000 0.0120 0.0080 0.018000000
#> 36  2005-03-02 12:56:30 0.001000000 0.00650000 0.0210 0.1820 0.203000000
#> 37  2005-04-04 11:51:00 0.005000000 0.04650000 0.0240 0.2450 0.269000000
#> 38  2005-05-02 10:46:00 0.004500000 0.03050000 0.0095 0.2285 0.238000000
#> 39  2005-06-06 10:09:00 0.003500000 0.02150000 0.0080 0.2805 0.288500000
#> 40  2005-08-09 09:29:30 0.002000000 0.03250000 0.0030 0.1420 0.145000000
#> 41  2005-09-08 09:26:30 0.001000000 0.08250000 0.0040     NA 0.007000000
#> 42  2005-10-03 09:26:00 0.001000000 0.04100000 0.0030 0.0120 0.011000000
#> 43  2005-11-08 11:11:00 0.002000000 0.00850000 0.0020 0.0675 0.069500000
#> 44  2005-11-29 13:46:00 0.001000000 0.00400000 0.0020     NA 0.007000000
#> 45  2006-01-03 12:02:00 0.001500000 0.01700000 0.0035 0.2880 0.291500000
#> 46  2006-02-01 11:17:30 0.002000000 0.02150000 0.0040 0.3015 0.305500000
#> 47  2006-03-01 12:17:00 0.001500000 0.03550000 0.0040 0.2475 0.251500000
#> 48  2006-04-04 11:28:00 0.002000000 0.08300000     NA     NA 0.211000000
#> 49  2006-05-02 11:11:00          NA         NA     NA     NA          NA
#> 50  2006-06-07 11:08:30          NA         NA     NA     NA          NA
#> 51  2006-07-05 09:29:00          NA         NA     NA     NA          NA
#> 52  2006-08-08 09:52:30 0.006000000 0.01000000     NA     NA 0.007000000
#> 53  2006-09-06 09:32:00 0.002000000 0.05050000     NA     NA 0.020000000
#> 54  2006-10-03 10:30:00 0.006500000 0.04000000     NA     NA 0.013500000
#> 55  2006-11-01 09:48:00 0.009000000 0.03000000     NA     NA 0.015500000
#> 56  2006-11-28 12:30:00 0.001000000 0.00400000     NA     NA 0.007000000
#> 57  2007-01-09 12:35:00 0.006500000 0.04650000     NA     NA 0.166000000
#> 58  2007-02-05 10:35:00 0.002000000 0.00950000     NA     NA 0.155500000
#> 59  2007-03-06 11:31:40 0.002333333 0.02766667     NA     NA 0.286000000
#> 60  2007-04-09 09:44:00 0.002000000 0.03300000     NA     NA 0.113000000
#> 61  2007-05-08 08:30:00 0.003000000 0.09100000     NA     NA 0.010000000
#> 62  2007-06-05 08:11:00 0.009000000 0.07500000     NA     NA 0.052000000
#> 63  2007-07-10 08:00:00 0.002000000 0.05900000     NA     NA 0.018000000
#> 64  2007-08-07 08:12:00 0.002000000 0.06700000     NA     NA 0.182000000
#> 65  2007-09-10 10:23:00 0.006000000 0.01100000     NA     NA 0.002000000
#> 66  2007-10-09 09:08:00 0.006000000 0.02700000     NA     NA 0.014000000
#> 67  2007-11-05 13:50:00 0.002000000 0.01600000     NA     NA 0.003000000
#> 68  2007-12-04 12:30:00 0.004000000 0.02100000     NA     NA 0.012000000
#> 69  2008-01-06 11:44:00 0.003000000 0.01700000     NA     NA 0.010000000
#> 70  2008-02-11 11:42:00 0.002000000 0.03033333     NA     NA 0.137333333
#> 71  2008-03-11 10:41:00 0.002000000 0.06400000     NA     NA 0.143000000
#> 72  2008-04-21 08:36:00 0.002000000 0.02400000     NA     NA 0.178000000
#> 73  2008-05-13 11:50:00 0.002000000 0.04500000     NA     NA 0.055000000
#> 74  2008-06-09 08:36:00 0.003000000 0.04800000     NA     NA 0.021000000
#> 75  2008-07-08 13:26:00 0.003000000 0.03800000     NA     NA 0.017000000
#> 76  2008-08-05 07:56:00 0.005000000 0.11000000     NA     NA 0.008000000
#> 77  2008-09-08 08:00:00 0.005000000 0.09700000     NA     NA 0.034000000
#> 78  2008-10-07 11:57:00 0.003000000 0.01600000     NA     NA 0.006200000
#> 79  2008-11-04 08:35:00 0.004000000 0.14500000     NA     NA 0.006200000
#> 80  2008-12-03 11:55:00 0.002000000 0.02600000     NA     NA 0.096000000
#> 81  2009-01-13 10:29:00 0.002000000 0.03400000     NA     NA 0.241000000
#> 82  2009-02-10 13:03:00 0.002000000 0.01170000     NA     NA 0.013000000
#> 83  2009-03-10 11:27:00 0.002000000 0.02033333     NA     NA 0.232666667
#> 84  2009-04-15 11:14:00 0.018000000 0.04900000     NA     NA 0.193000000
#> 85  2009-05-05 13:35:00 0.003000000 0.06300000     NA     NA 0.193000000
#> 86  2009-06-09 08:10:00 0.004000000 0.11000000     NA     NA 0.117000000
#> 87  2009-07-14 12:08:00 0.001000000 0.04400000     NA     NA 0.054000000
#> 88  2009-08-04 12:03:00 0.001000000 0.01930000     NA     NA 0.034000000
#> 89  2009-09-08 11:56:00 0.005333333 0.08900000     NA     NA 0.008666667
#> 90  2009-10-06 09:30:00 0.009000000 0.13000000     NA     NA 0.120000000
#> 91  2009-11-03 12:40:00 0.002000000 0.02600000     NA     NA 0.041000000
#> 92  2009-12-08 11:30:00 0.007000000 0.01930000     NA     NA 0.858000000
#> 93  2010-01-12 12:28:00 0.006000000 0.02800000     NA     NA 0.400000000
#> 94  2010-02-02 13:31:00 0.007000000 0.05000000     NA     NA 0.369000000
#> 95  2010-03-04 14:35:00 0.002000000 0.02300000     NA     NA 0.430000000
#> 96  2010-04-06 08:31:00 0.006000000 0.05200000     NA     NA 0.227000000
#> 97  2010-05-05 08:31:00 0.004000000 0.07300000     NA     NA 0.156000000
#> 98  2010-06-08 14:03:00 0.002000000 0.01770000     NA     NA 0.025000000
#> 99  2010-07-13 12:38:00 0.002000000 0.01770000     NA     NA 0.022000000
#> 100 2010-08-04 07:24:00 0.002000000 0.04400000     NA     NA 0.034000000
#> 101 2010-09-08 11:40:00 0.002000000 0.04300000     NA     NA 0.002100000
#> 102 2010-10-06 10:00:40 0.002000000 0.02023333     NA     NA 0.068000000
#> 103 2010-11-02 09:53:00 0.009000000 0.04800000     NA     NA 0.018000000
#> 104 2010-12-14 12:26:00 0.002000000 0.05000000     NA     NA 0.246000000
#> 105 2011-01-11 11:28:00 0.003000000 0.04300000     NA     NA 0.003000000
#> 106 2011-02-09 12:56:20 0.004333333 0.01500000     NA     NA 0.086666667
#> 107 2011-03-02 10:36:00 0.004000000 0.01500000     NA     NA 0.047000000
#> 108 2011-04-06 11:29:00 0.003000000 0.04200000     NA     NA 0.245000000
#> 109 2011-05-11 10:38:00 0.003000000 0.03700000     NA     NA 0.017000000
#> 110 2011-06-07 09:54:00 0.003000000 0.02200000     NA     NA 0.010000000
#> 111 2011-07-13 09:43:00 0.003000000 0.03600000     NA     NA 0.018000000
#> 112 2011-08-03 09:30:00 0.003000000 0.06600000     NA     NA 0.030000000
#> 113 2011-09-07 10:12:00 0.003000000 0.08400000     NA     NA 0.065000000
#> 114 2011-10-05 11:05:00 0.003000000 0.06200000     NA     NA 0.003000000
#> 115 2011-11-02 09:34:00 0.003000000 0.01800000     NA     NA 0.004000000
#> 116 2011-12-13 11:15:00 0.003000000 0.03200000     NA     NA 0.007000000
#> 117 2012-01-10 10:20:00 0.003000000 0.03000000     NA     NA 0.002400000
#> 118 2012-02-07 11:41:00 0.005000000 0.01900000     NA     NA 0.081000000
#> 119 2012-03-05 11:51:00 0.003000000 0.04100000     NA     NA 0.510000000
#> 120 2012-04-04 10:30:00 0.003000000 0.04300000     NA     NA 0.154000000
#> 121 2012-05-09 10:15:40 0.003000000 0.03033333     NA     NA 0.006666667
#> 122 2012-06-05 08:30:00 0.003000000 0.04000000     NA     NA 0.042000000
#> 123 2012-07-03 09:59:20 0.004333333 0.07633333     NA     NA 0.058000000
#> 124 2012-08-07 09:53:00 0.003000000 0.05000000     NA     NA 0.049000000
#> 125 2012-09-05 10:56:00 0.003000000 0.02600000     NA     NA 0.007000000
#> 126 2012-10-02 09:27:00 0.003000000 0.03600000     NA     NA 0.012333333
#> 127 2012-11-06 10:30:00 0.003000000 0.07000000     NA     NA 0.006000000
#> 128 2012-11-26 11:39:00 0.003000000 0.04100000     NA     NA 0.004000000
#> 129 2013-01-08 10:54:00 0.003000000 0.01300000     NA     NA 0.004000000
#> 130 2013-02-05 10:23:00 0.008000000 0.01400000     NA     NA 0.002000000
#> 131 2013-03-04 15:47:20 0.012666667 0.06666667     NA     NA 0.236000000
#> 132 2013-04-02 10:18:00 0.003000000 0.06300000     NA     NA 0.276000000
#> 133 2013-05-08 10:07:00 0.003000000 0.06500000     NA     NA 0.100000000
#> 134 2013-06-03 09:12:20 0.003000000 0.05366667     NA     NA 0.002000000
#> 135 2013-07-09 08:56:00 0.004000000 0.06500000     NA     NA 0.036000000
#> 136 2013-08-12 09:22:00 0.004000000 0.01900000     NA     NA 0.019000000
#> 137 2013-09-04 08:13:00 0.006000000 0.10000000     NA     NA 0.150000000
#> 138 2013-10-08 10:07:00 0.005000000 0.03300000     NA     NA 0.069000000
#> 139 2013-10-29 09:30:00 0.004000000 0.01200000     NA     NA 0.009000000
#> 140 2013-12-03 11:07:00 0.005000000 0.01000000     NA     NA 0.150000000
#>        chla_n
#> 1    1.800000
#> 2    1.200000
#> 3    3.400000
#> 4    3.350000
#> 5    7.800000
#> 6    0.800000
#> 7    2.350000
#> 8    3.700000
#> 9    2.350000
#> 10  13.243000
#> 11  14.952000
#> 12   7.903500
#> 13   2.990000
#> 14   6.087500
#> 15   4.485500
#> 16   8.117000
#> 17   3.845000
#> 18   8.544000
#> 19   1.282000
#> 20   1.495500
#> 21   2.136000
#> 22   1.050000
#> 23   2.600000
#> 24   1.900000
#> 25   2.600000
#> 26   3.600000
#> 27   2.100000
#> 28   4.150000
#> 29   6.100000
#> 30  17.300000
#> 31   4.300000
#> 32   0.650000
#> 33   0.650000
#> 34   1.600000
#> 35   1.495000
#> 36   0.740000
#> 37   1.090000
#> 38   1.825000
#> 39   1.705000
#> 40   1.495000
#> 41   2.560000
#> 42   2.560000
#> 43   1.065000
#> 44   1.280000
#> 45   4.740000
#> 46   3.885000
#> 47   1.960000
#> 48   4.710000
#> 49         NA
#> 50         NA
#> 51         NA
#> 52  10.425000
#> 53   7.350000
#> 54   2.520000
#> 55   1.560000
#> 56   4.935000
#> 57   4.450000
#> 58   3.050000
#> 59   2.900000
#> 60   3.800000
#> 61   8.500000
#> 62   2.900000
#> 63   6.500000
#> 64   8.000000
#> 65  13.600000
#> 66   4.200000
#> 67   4.700000
#> 68   2.700000
#> 69   0.400000
#> 70   3.066667
#> 71   6.500000
#> 72   4.900000
#> 73  14.500000
#> 74   7.500000
#> 75   9.600000
#> 76  15.600000
#> 77   5.800000
#> 78   5.800000
#> 79   2.900000
#> 80   5.400000
#> 81   6.900000
#> 82   8.000000
#> 83   5.800000
#> 84   2.400000
#> 85   9.400000
#> 86   6.000000
#> 87  11.800000
#> 88  19.700000
#> 89   7.633333
#> 90   5.800000
#> 91   6.900000
#> 92   2.700000
#> 93   1.800000
#> 94   4.700000
#> 95   5.100000
#> 96   4.500000
#> 97   5.100000
#> 98  13.600000
#> 99  13.400000
#> 100 11.600000
#> 101 11.600000
#> 102  4.533333
#> 103  3.800000
#> 104  6.500000
#> 105  3.800000
#> 106  8.566667
#> 107  4.700000
#> 108  9.400000
#> 109 10.500000
#> 110 12.300000
#> 111 14.700000
#> 112 14.500000
#> 113 10.000000
#> 114  6.200000
#> 115  6.000000
#> 116  3.400000
#> 117  4.020000
#> 118  6.700000
#> 119  8.040000
#> 120  5.810000
#> 121  9.013333
#> 122  9.380000
#> 123  6.480000
#> 124  6.700000
#> 125  8.710000
#> 126 13.183333
#> 127  3.130000
#> 128  2.460000
#> 129  2.230000
#> 130  4.020000
#> 131  2.156667
#> 132 10.500000
#> 133  5.590000
#> 134  5.810000
#> 135  6.800000
#> 136 11.000000
#> 137  4.400000
#> 138  7.400000
#> 139  2.700000
#> 140  5.900000

# use different function to aggregate replicates
func <- function(x) max(x, na.rm = TRUE)
rem_reps(swmp1, FUN = func)
#> Warning: QAQC columns present, removed from output
#>           datetimestamp  po4f   nh4f  no2f  no3f  no23f chla_n
#> 1   2002-04-02 11:55:30 0.004 0.0290 0.002 0.048 0.0500  1.800
#> 2   2002-04-30 11:15:00 0.014 0.1380 0.005 0.115 0.1200  1.200
#> 3   2002-06-04 11:03:00 0.006 0.0490 0.002 0.024 0.0260  3.400
#> 4   2002-07-02 09:54:00 0.017 0.0930 0.002    NA 0.0400  3.700
#> 5   2002-08-06 12:13:00 0.011 0.0400 0.003 0.036 0.0390  7.800
#> 6   2002-09-10 09:00:00 0.026 0.0390 0.003 0.013 0.0160  0.800
#> 7   2002-10-08 08:56:00 0.010 0.0370 0.009 0.020 0.0290  2.600
#> 8   2002-11-04 10:05:30 0.007 0.0620 0.006 0.101 0.1060  4.000
#> 9   2002-12-02 12:10:30 0.011 0.0650 0.004 0.110 0.1140  2.600
#> 10  2003-01-06 10:31:00 0.014 0.3600 0.003 0.239 0.2420 13.670
#> 11  2003-02-03 11:11:00 0.004 0.0400 0.002    NA 0.1060 15.379
#> 12  2003-03-03 10:21:00 0.016 0.0640 0.007 0.131 0.1360  8.117
#> 13  2003-04-08 08:30:00 0.025 0.1590 0.005 0.174 0.1790  2.990
#> 14  2003-05-06 09:49:00 0.004 0.0600 0.007 0.246 0.2530  7.049
#> 15  2003-06-02 10:16:00 0.004 0.0610 0.009 0.223 0.2320  4.699
#> 16  2003-07-08 12:47:00 0.001 0.0350 0.002 0.052 0.0540  8.117
#> 17  2003-08-04 10:18:30 0.001 0.0360 0.003 0.312 0.3150  3.845
#> 18  2003-09-09 12:12:00 0.003 0.0140 0.005 0.141 0.1460  8.544
#> 19  2003-10-07 12:48:30 0.005 0.1440 0.002 0.030 0.0320  1.282
#> 20  2003-11-05 09:38:30 0.004 0.0170 0.002 0.026 0.0290  1.709
#> 21  2003-12-08 10:57:30 0.002 0.0040 0.002 0.185 0.1870  2.136
#> 22  2004-01-08 10:51:00 0.001 0.0080 0.002    NA 0.0190  1.300
#> 23  2004-02-09 11:57:30 0.001 0.0060 0.002 0.082 0.0840  2.600
#> 24  2004-03-09 11:11:00 0.002 0.0340 0.003 0.286 0.2890  2.100
#> 25  2004-04-05 09:52:30 0.001 0.0210 0.002 0.097 0.1030  2.600
#> 26  2004-05-04 10:06:00 0.008 0.0870 0.005 0.090 0.0940  3.800
#> 27  2004-06-07 09:17:00 0.008 0.0600 0.004 0.038 0.0420  2.100
#> 28  2004-07-06 08:57:30 0.001 0.0600 0.006 0.062 0.0680  4.500
#> 29  2004-08-02 12:41:00 0.001 0.0140 0.002    NA 0.0260  6.400
#> 30  2004-09-08 14:07:30 0.004 0.0480 0.013 0.118 0.1310 32.000
#> 31  2004-10-04 10:08:30    NA 0.0500 0.013 0.044 0.0550  4.300
#> 32  2004-11-08 10:16:30 0.002 0.0200 0.013 0.182 0.1930  0.800
#> 33  2004-11-30 12:02:00 0.009 0.0200 0.005 0.039 0.0440  0.800
#> 34  2005-01-04 10:47:30 0.004 0.0310 0.002 0.014 0.0190  1.920
#> 35  2005-02-07 11:18:00 0.001 0.0190 0.012 0.008 0.0200  1.710
#> 36  2005-03-02 12:56:30 0.001 0.0090 0.024 0.184 0.2080  0.980
#> 37  2005-04-04 11:51:00 0.005 0.0470 0.024 0.252 0.2760  1.220
#> 38  2005-05-02 10:46:00 0.005 0.0310 0.010 0.235 0.2440  1.920
#> 39  2005-06-06 10:09:00 0.004 0.0220 0.008 0.282 0.2900  2.130
#> 40  2005-08-09 09:29:30 0.002 0.0340 0.003 0.147 0.1500  1.710
#> 41  2005-09-08 09:26:30 0.001 0.0860 0.004    NA 0.0070  2.560
#> 42  2005-10-03 09:26:00 0.001 0.0420 0.003 0.012 0.0150  2.560
#> 43  2005-11-08 11:11:00 0.002 0.0090 0.002 0.099 0.1010  1.280
#> 44  2005-11-29 13:46:00 0.001 0.0040 0.002    NA 0.0070  1.280
#> 45  2006-01-03 12:02:00 0.002 0.0180 0.004 0.292 0.2960  5.250
#> 46  2006-02-01 11:17:30 0.002 0.0220 0.004 0.313 0.3170  3.930
#> 47  2006-03-01 12:17:00 0.002 0.0360 0.004 0.249 0.2530  2.130
#> 48  2006-04-04 11:28:00 0.002 0.0830    NA    NA 0.2240  4.740
#> 49  2006-05-02 11:11:00    NA     NA    NA    NA     NA     NA
#> 50  2006-06-07 11:08:30    NA     NA    NA    NA     NA     NA
#> 51  2006-07-05 09:29:00    NA     NA    NA    NA     NA     NA
#> 52  2006-08-08 09:52:30 0.006 0.0100    NA    NA 0.0070 10.550
#> 53  2006-09-06 09:32:00 0.002 0.0510    NA    NA 0.0210  7.820
#> 54  2006-10-03 10:30:00 0.007 0.0460    NA    NA 0.0150  2.730
#> 55  2006-11-01 09:48:00 0.009 0.0300    NA    NA 0.0180  1.620
#> 56  2006-11-28 12:30:00 0.001 0.0040    NA    NA 0.0070  4.960
#> 57  2007-01-09 12:35:00 0.007 0.0480    NA    NA 0.1660  4.700
#> 58  2007-02-05 10:35:00 0.002 0.0110    NA    NA 0.1590  4.300
#> 59  2007-03-06 11:31:40 0.003 0.0300    NA    NA 0.2900  3.100
#> 60  2007-04-09 09:44:00 0.002 0.0330    NA    NA 0.1130  3.800
#> 61  2007-05-08 08:30:00 0.003 0.0910    NA    NA 0.0100  8.500
#> 62  2007-06-05 08:11:00 0.009 0.0750    NA    NA 0.0520  2.900
#> 63  2007-07-10 08:00:00 0.002 0.0590    NA    NA 0.0180  6.500
#> 64  2007-08-07 08:12:00 0.002 0.0670    NA    NA 0.1820  8.000
#> 65  2007-09-10 10:23:00 0.006 0.0110    NA    NA 0.0020 13.600
#> 66  2007-10-09 09:08:00 0.006 0.0270    NA    NA 0.0140  4.200
#> 67  2007-11-05 13:50:00 0.002 0.0160    NA    NA 0.0030  4.700
#> 68  2007-12-04 12:30:00 0.004 0.0210    NA    NA 0.0120  2.700
#> 69  2008-01-06 11:44:00 0.003 0.0170    NA    NA 0.0100  0.400
#> 70  2008-02-11 11:42:00 0.002 0.0310    NA    NA 0.1400  3.400
#> 71  2008-03-11 10:41:00 0.002 0.0640    NA    NA 0.1430  6.500
#> 72  2008-04-21 08:36:00 0.002 0.0240    NA    NA 0.1780  4.900
#> 73  2008-05-13 11:50:00 0.002 0.0450    NA    NA 0.0550 14.500
#> 74  2008-06-09 08:36:00 0.003 0.0480    NA    NA 0.0210  7.500
#> 75  2008-07-08 13:26:00 0.003 0.0380    NA    NA 0.0170  9.600
#> 76  2008-08-05 07:56:00 0.005 0.1100    NA    NA 0.0080 15.600
#> 77  2008-09-08 08:00:00 0.005 0.0970    NA    NA 0.0340  5.800
#> 78  2008-10-07 11:57:00 0.003 0.0160    NA    NA 0.0062  5.800
#> 79  2008-11-04 08:35:00 0.004 0.1450    NA    NA 0.0062  2.900
#> 80  2008-12-03 11:55:00 0.002 0.0260    NA    NA 0.0960  5.400
#> 81  2009-01-13 10:29:00 0.002 0.0340    NA    NA 0.2410  6.900
#> 82  2009-02-10 13:03:00 0.002 0.0117    NA    NA 0.0130  8.000
#> 83  2009-03-10 11:27:00 0.002 0.0230    NA    NA 0.2370  6.000
#> 84  2009-04-15 11:14:00 0.018 0.0490    NA    NA 0.1930  2.400
#> 85  2009-05-05 13:35:00 0.003 0.0630    NA    NA 0.1930  9.400
#> 86  2009-06-09 08:10:00 0.004 0.1100    NA    NA 0.1170  6.000
#> 87  2009-07-14 12:08:00 0.001 0.0440    NA    NA 0.0540 11.800
#> 88  2009-08-04 12:03:00 0.001 0.0193    NA    NA 0.0340 19.700
#> 89  2009-09-08 11:56:00 0.006 0.1290    NA    NA 0.0110  8.300
#> 90  2009-10-06 09:30:00 0.009 0.1300    NA    NA 0.1200  5.800
#> 91  2009-11-03 12:40:00 0.002 0.0260    NA    NA 0.0410  6.900
#> 92  2009-12-08 11:30:00 0.007 0.0193    NA    NA 0.8580  2.700
#> 93  2010-01-12 12:28:00 0.006 0.0280    NA    NA 0.4000  1.800
#> 94  2010-02-02 13:31:00 0.007 0.0500    NA    NA 0.3690  4.700
#> 95  2010-03-04 14:35:00 0.002 0.0230    NA    NA 0.4300  5.100
#> 96  2010-04-06 08:31:00 0.006 0.0520    NA    NA 0.2270  4.500
#> 97  2010-05-05 08:31:00 0.004 0.0730    NA    NA 0.1560  5.100
#> 98  2010-06-08 14:03:00 0.002 0.0177    NA    NA 0.0250 13.600
#> 99  2010-07-13 12:38:00 0.002 0.0177    NA    NA 0.0220 13.400
#> 100 2010-08-04 07:24:00 0.002 0.0440    NA    NA 0.0340 11.600
#> 101 2010-09-08 11:40:00 0.002 0.0430    NA    NA 0.0021 11.600
#> 102 2010-10-06 10:00:40 0.002 0.0220    NA    NA 0.0730  4.900
#> 103 2010-11-02 09:53:00 0.009 0.0480    NA    NA 0.0180  3.800
#> 104 2010-12-14 12:26:00 0.002 0.0500    NA    NA 0.2460  6.500
#> 105 2011-01-11 11:28:00 0.003 0.0430    NA    NA 0.0030  3.800
#> 106 2011-02-09 12:56:20 0.005 0.0150    NA    NA 0.0950  8.900
#> 107 2011-03-02 10:36:00 0.004 0.0150    NA    NA 0.0470  4.700
#> 108 2011-04-06 11:29:00 0.003 0.0420    NA    NA 0.2450  9.400
#> 109 2011-05-11 10:38:00 0.003 0.0370    NA    NA 0.0170 10.500
#> 110 2011-06-07 09:54:00 0.003 0.0220    NA    NA 0.0100 12.300
#> 111 2011-07-13 09:43:00 0.003 0.0360    NA    NA 0.0180 14.700
#> 112 2011-08-03 09:30:00 0.003 0.0660    NA    NA 0.0300 14.500
#> 113 2011-09-07 10:12:00 0.003 0.0840    NA    NA 0.0650 10.000
#> 114 2011-10-05 11:05:00 0.003 0.0620    NA    NA 0.0030  6.200
#> 115 2011-11-02 09:34:00 0.003 0.0180    NA    NA 0.0040  6.000
#> 116 2011-12-13 11:15:00 0.003 0.0320    NA    NA 0.0070  3.400
#> 117 2012-01-10 10:20:00 0.003 0.0300    NA    NA 0.0024  4.020
#> 118 2012-02-07 11:41:00 0.005 0.0190    NA    NA 0.0810  6.700
#> 119 2012-03-05 11:51:00 0.003 0.0410    NA    NA 0.5100  8.040
#> 120 2012-04-04 10:30:00 0.003 0.0430    NA    NA 0.1540  5.810
#> 121 2012-05-09 10:15:40 0.003 0.0530    NA    NA 0.0090  9.160
#> 122 2012-06-05 08:30:00 0.003 0.0400    NA    NA 0.0420  9.380
#> 123 2012-07-03 09:59:20 0.005 0.0940    NA    NA 0.0600  8.040
#> 124 2012-08-07 09:53:00 0.003 0.0500    NA    NA 0.0490  6.700
#> 125 2012-09-05 10:56:00 0.003 0.0260    NA    NA 0.0070  8.710
#> 126 2012-10-02 09:27:00 0.003 0.0420    NA    NA 0.0150 13.850
#> 127 2012-11-06 10:30:00 0.003 0.0700    NA    NA 0.0060  3.130
#> 128 2012-11-26 11:39:00 0.003 0.0410    NA    NA 0.0040  2.460
#> 129 2013-01-08 10:54:00 0.003 0.0130    NA    NA 0.0040  2.230
#> 130 2013-02-05 10:23:00 0.008 0.0140    NA    NA 0.0020  4.020
#> 131 2013-03-04 15:47:20 0.016 0.0720    NA    NA 0.2480  2.680
#> 132 2013-04-02 10:18:00 0.003 0.0630    NA    NA 0.2760 10.500
#> 133 2013-05-08 10:07:00 0.003 0.0650    NA    NA 0.1000  5.590
#> 134 2013-06-03 09:12:20 0.003 0.0630    NA    NA 0.0020  5.810
#> 135 2013-07-09 08:56:00 0.004 0.0650    NA    NA 0.0360  6.800
#> 136 2013-08-12 09:22:00 0.004 0.0190    NA    NA 0.0190 11.000
#> 137 2013-09-04 08:13:00 0.006 0.1000    NA    NA 0.1500  4.400
#> 138 2013-10-08 10:07:00 0.005 0.0330    NA    NA 0.0690  7.400
#> 139 2013-10-29 09:30:00 0.004 0.0120    NA    NA 0.0090  2.700
#> 140 2013-12-03 11:07:00 0.005 0.0100    NA    NA 0.1500  5.900